+++ src/test/analyzer_test.ts
@@ -29,7 +29,6 @@ import {JavaScriptDocument} from '../javascript/javascript-document';
 import {Document, Import, ScannedImport, ScannedInlineDocument} from '../model/model';
 import {FSUrlLoader} from '../url-loader/fs-url-loader';
 import {UrlLoader} from '../url-loader/url-loader';
-import {UrlResolver} from '../url-loader/url-resolver';
 import {Deferred} from '../utils';
 import {Severity} from '../warning/warning';
 import {CodeUnderliner} from './test-utils';
@@ -41,28 +40,22 @@ import stripIndent = require('strip-indent');
 use(chaiSubset);
 use(chaiAsPromised);
 
-class TestUrlResolver implements UrlResolver {
-  canResolve(url: string) {
-    return (url === 'test.com/test.html');
-  }
-
-  resolve(url: string) {
-    return (url === 'test.com/test.html') ? '/static/html-parse-target.html' :
-                                            url;
-  }
-}
-
 suite('Analyzer', () => {
   let analyzer: Analyzer;
   let underliner: CodeUnderliner;
 
   setup(() => {
     const urlLoader = new FSUrlLoader(__dirname);
-    analyzer = new Analyzer({
-      urlLoader,
-      urlResolver: new TestUrlResolver(),
-    });
+    analyzer = new Analyzer({urlLoader});
     underliner = new CodeUnderliner(urlLoader);
+  });
+
+  test('canResolveUrl defaults to not resolving external urls', () => {
+    assert.isTrue(analyzer.canResolveUrl('/path'), '/path');
+    assert.isTrue(analyzer.canResolveUrl('../path'), '../path');
+    assert.isFalse(analyzer.canResolveUrl('http://host'), 'http://host');
+    assert.isFalse(
+        analyzer.canResolveUrl('http://host/path'), 'http://host/path');
   });
 
   suite('analyze()', () => {